requirements.txt
pytest
pyyaml
requests
requests_mock
rtmidi
yapf
